# lib/protocol: Faster Luhn algorithm and better testing