### Purpose

The previous implementation of the Luhn algorithm was very generic; its tests didn't cover the actual alphabet for device IDs. This new specialized version is shorter and has tests that cover the actual base-32 Luhn used. It's also almost 4× faster:

```
name         old time/op    new time/op     delta
Luhnify-8      1.00µs ± 1%     0.28µs ± 4%   -72.38%  (p=0.000 n=9+10)
Unluhnify-8     992ns ± 2%      274ns ± 1%   -72.39%  (p=0.000 n=10+9)
```

The unused luhnValidate function is gone.