@AudriusButkevicius 2020-03-29
So I think our luhn is not actually luhn, and it has an implementation bug, so we probably need to retain that.