@greatroar 2020-03-29
I spotted the comment. This PR retains the current behavior, broken or not, and the test asserts that (try running it on master).