@calmh 2020-03-29
This may well be better, but there is zero performance requirements on this code and the existing one has worked untouched for years. As such this feels more like churn than an improvement.